Graphical interface working again
