+++ src/actions/index.js
@@ -1,4 +1,4 @@
-import INITIALIZE from './initialize'
+import INITIALIZE from './initialize/index'
 import CHANGE_FILTER from './changeFilter'
 import RETRIEVE_FILTERED_STOCKS from './retrieveFilteredStocks'
 import CLEAR_FILTERS from './clearFilters'
